fix error converting controller method lcfirst
If controller is not defined in the URI, then an error occurs if an
attempt is made to convert the missing segment to 'lower case first'.
This fix amends that by converting the resolved controller method name
to 'lcfirst' even if it comes from the protected class property as
default value.